pyproject.toml
[tool.poetry]
name = "habit-tracker-3"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.4.0"
django-celery-beat = "^2.7.0"
redis = "^5.2.1"
celery = "^5.4.0"
eventlet = "^0.38.2"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"
coverage = "^7.6.10"
django-filter = "^24.3"
pillow = "^11.1.0"
requests = "^2.32.3"
pytest = "^8.3.4"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
